# contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/assets/{id}/meta": {
      "get": {
        "operationId": "AssetsController_assetMeta",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assets"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/assets/{id}/file": {
      "get": {
        "operationId": "AssetsController_assetFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assets"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/playlists/{id}/manifest": {
      "get": {
        "operationId": "PlaylistsController_manifest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "playlists"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/devices/{id}/state": {
      "get": {
        "operationId": "DevicesController_getState",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "devices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/monitoring/devices/{deviceId}/preview": {
      "get": {
        "operationId": "MonitoringController_devicePreviewImage",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "monitoring"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/device/telemetry": {
      "post": {
        "operationId": "DeviceApiController_postTelemetry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/playback-events": {
      "post": {
        "operationId": "DeviceApiController_postPlaybackEvents",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackEventBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/preview": {
      "post": {
        "operationId": "DeviceApiController_postPreview",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/commands/pending": {
      "get": {
        "operationId": "DeviceApiController_pendingCommands",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/commands/{commandId}/ack": {
      "post": {
        "operationId": "DeviceApiController_ackCommand",
        "parameters": [
          {
            "name": "commandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCommandAckDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/releases/latest": {
      "get": {
        "operationId": "DeviceApiController_latestRelease",
        "parameters": [
          {
            "name": "product",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/heartbeat": {
      "post": {
        "operationId": "DeviceApiController_heartbeat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/state": {
      "get": {
        "operationId": "DeviceApiController_state",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/playlists/{playlistId}/manifest": {
      "get": {
        "operationId": "DeviceApiController_playlistManifest",
        "parameters": [
          {
            "name": "playlistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/assets/{assetId}/meta": {
      "get": {
        "operationId": "DeviceApiController_assetMeta",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
    "/device/assets/{assetId}/file": {
      "get": {
        "operationId": "DeviceApiController_assetFile",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ],
        "security": [
          {
            "device-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EasySignage API",
    "description": "REST do CMS (JWT), rotas públicas de pareamento e API do player (Bearer de device). Especificação evolutiva — alinhar com `digital_signage_arquitetura_roadmap.md`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "device-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TelemetryBatchDto": {
        "type": "object",
        "properties": {}
      },
      "PlaybackEventBatchDto": {
        "type": "object",
        "properties": {}
      },
      "DeviceCommandAckDto": {
        "type": "object",
        "properties": {}
      },
      "HeartbeatDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}